Feat: move `calculatePercentageReputation` to the utils
